Feat: Complete reservation flow, switch to file-based H2, add returns module

+++ src/main/java/com/autodrive/backend/repository/ReturnReportRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReturnReportRepository extends JpaRepository<ReturnReport, Integer> {
+public interface ReturnReportRepository extends JpaRepository<ReturnReport, Long> {
     
 }
